Parse `limit` and `offset` as UInt64 in SQL (#185)
This avoids casting errors on the ClickHouse side.
In-line with the max value for `page`.

+++ src/sql/nft_holders/evm.sql
@@ -41,5 +41,5 @@ combined AS (
 )
 SELECT *
 FROM combined
-LIMIT {limit:int}
-OFFSET {offset:int}
+LIMIT {limit:UInt64}
+OFFSET {offset:UInt64}
